[_Cilk_for] Modified _Cilk_for loops such that, if the loop condition is simple, then the loop limit is evaluated before the loop.
For example, consider the following _Cilk_for loop:

_Cilk_for (int i = X; i < Y; ++i)

This loop is transformed such that Y is evaluated once before the
loop executes.  In effect, this loop is treated as syntactic sugar
for the following loop:

_Cilk_for (int i = X, __end = Y; i < __end; ++i)

More generally, Y can be any expression that does not use a
variable declared in the initialization.