[Cilk] Added preliminary support for _Cilk_spawn assignments and declarations.
This commit adds support for the following statements and expressions:

var = _Cilk_spawn function-call
vardecl = _Cilk_spawn function-call

This support should be easily extensible to support spawned compound
assignments by means of Cilk inlets.

To support these statements and expressions, this commit adds a CilkSpawnExpr
expression type, which has a single child subexpression.  Although
_Cilk_spawn cannot be applied to any expression, this approach of adding a
general CilkSpawnExpr seems to be the simplest for adapting the parser to
handle the cases we care about.  Additional semantic checks should be added
later to ensure that a CilkSpawnExpr appears only where it makes semantic
sense.

_Cilk_spawn assignments and declarations differ from ordinary assignments and
declarations in their order of evaluation.  For a typical assignment or
declaration,

E1 @= E2

C++17 rules dictate that E2 is evaluated before E1.  When E2 is preceded by
a _Cilk_spawn, however, we no longer guarantee that E2 is evaluated before
E1.  (This commit, in fact, ensures that E1 is evaluated before E2 if E2 is
spawned.)  In effect, the _Cilk_spawn weakens the evaluation-order
guarantees of the spawned expression and code evaluated afterwards.

This commit also improves the handling of evaluated expressions around Cilk
constructs.  These improvements include fixes to the evaluation order of
arguments to a spawned function, creation of temporary storage for spawned
function calls, and processing of exception-handling code created for
spawned expressions.  This commit also reorganizes some Cilk-specific code
into separate files.